Release v3.2.1

+++ cookbooks/arcgis-geoevent/recipes/default.rb
@@ -33,7 +33,9 @@ arcgis_geoevent_geoevent 'Authorize ArcGIS GeoEvent Server' do
 end
 
 file node['arcgis']['geoevent']['cached_authorization_file'] do
-  content File.open(node['arcgis']['geoevent']['authorization_file'], 'rb') { |file| file.read }
+  if ::File.exists?(node['arcgis']['geoevent']['authorization_file'])
+    content File.open(node['arcgis']['geoevent']['authorization_file'], 'rb') { |file| file.read }
+  end
   sensitive true
   subscribes :create, 'arcgis_geoevent_geoevent[Authorize ArcGIS GeoEvent Server]', :immediately
   action :nothing
@@ -46,6 +48,23 @@ if node['platform'] == 'windows'
     only_if { Utils.product_installed?(node['arcgis']['geoevent']['product_code']) }
     action :update_account #TODO: trigger update_account only when arcgis user account is updated
   end
+end
+
+arcgis_geoevent_geoevent 'Unpack ArcGIS GeoEvent Server Setup' do
+  setup_archive node['arcgis']['geoevent']['setup_archive']
+  setups_repo node['arcgis']['repository']['setups']
+  run_as_user node['arcgis']['run_as_user']
+  only_if { ::File.exist?(node['arcgis']['geoevent']['setup_archive']) &&
+            !::File.exist?(node['arcgis']['geoevent']['setup']) }
+  if node['platform'] == 'windows'
+    not_if { Utils.product_installed?(node['arcgis']['geoevent']['product_code']) }
+  else
+    not_if { EsriProperties.product_installed?(node['arcgis']['run_as_user'],
+                                               node['hostname'],
+                                               node['arcgis']['version'],
+                                               :ArcGISGeoEvent) }
+  end
+  action :unpack
 end
 
 arcgis_geoevent_geoevent 'Setup ArcGIS GeoEvent Server' do
@@ -69,6 +88,37 @@ arcgis_geoevent_geoevent 'Configure ArcGISGeoEvent service' do
   install_dir node['arcgis']['server']['install_dir']
   action :configure_autostart
   only_if { node['arcgis']['geoevent']['configure_autostart'] }
+end
+
+# Sometimes rabbitmq file is not copied to the user's AppData folder 
+# by ArcGIS Server. Copy the file to work around the issue.
+execute 'Create RabbitMQ dir' do
+  command "mkdir \"C:\\Users\\#{node['arcgis']['run_as_user']}\\AppData\\Roaming\\RabbitMQ\""
+  user node['arcgis']['run_as_user']
+  password node['arcgis']['run_as_password']
+  returns [0, 1]
+  only_if { node['platform'] == 'windows' }
+end
+
+execute 'Copy rabbitmq file' do
+  command "copy \"#{node['arcgis']['server']['install_dir']}\\framework\\runtime\\rabbitmq\\etc\\rabbitmq\" " + 
+          "\"C:\\Users\\#{node['arcgis']['run_as_user']}\\AppData\\Roaming\\RabbitMQ\""
+  user node['arcgis']['run_as_user']
+  password node['arcgis']['run_as_password']
+  # ignore_failure true
+  only_if { node['platform'] == 'windows' }
+end
+
+# Remove everything under C:\ProgramData\Esri\GeoEvent-Gateway before starting GeoEvent.
+directory "#{ENV['ProgramData']}\\Esri\\GeoEvent-Gateway" do
+  recursive true
+  only_if { node['platform'] == 'windows' }
+  action :delete
+end
+
+directory "#{ENV['ProgramData']}\\Esri\\GeoEvent-Gateway" do
+  only_if { node['platform'] == 'windows' }
+  action :create
 end
 
 arcgis_geoevent_geoevent 'Start ArcGIS GeoEvent' do
